version 1.07 (commit generated from cpan)
Fixed some incorrect error messages in Streamer.xs. Updated CHANGES with 1.06
info which I forgot from that release. Fixed utf8 related error. Thanks to
Yitzchak for the patch. Provided a way to optionally install a 'DDS' alias
for the module so that its easier to use from the command line. Doesn't install
it by default, or via a normal CPAN install for namespace pollution reasons,
and because I hate modules that dont automatically install. If you do

  perl Makefile.PL DDS [STANDARD OPTS]

then it will be installed. To disable this once its enabled you can say

  perl Makefile.PL NODDS [STANDARD OPTS]