version 2.00 (commit generated from cpan)
*WARNING* Interface change!!

This release is interface incompatible with earlier releases.

The whole framework for overriding how objects and classes are represented
has been changed, hopefully, to be more intuitive and easier to use.

All of the previous Freeze related functionality has been replaced by the
new Freezer() method and interface. There is now support for class
authors to add a DDS_freeze() method to control serialization and
deserialization.

HashKeys has had an interface change and has been renamed to KeyOrder.
HashKeys() is still maintained but undocumented at this time.

Added more documentation.

Added support for Padwalker functionality that Robin Housten implemented.

Long quoted strings are now split to be easier to read.