version 2.03 (commit generated from cpan)
Fixed some issues with dmping proxy objects with circular links.

Started integrating subversion into the perl build process to prevent
stuff like releasing unchecked in code, and to include the subversion
build number in the distribution file name.

Reversed the order of the change file so that newest stuff is at top.