Fix [RT #41835] read-only objects (from XS in this case) break things
We try to rebless items into a safe namespace so we cant be fooled by
their overloaded behavior. However this does not work with read only
objects, which apparently JSON::XS::Boolean creates (why it creates this
and not a true boolean I do not understand).

For now this adds JSON::XS to the dependency list. I will watch for
bugs related to requiring this module.